requirements.txt
requests~=2.25.1
asyncio~=3.4.3
aiohttp~=3.7.4.post0
beautifulsoup4~=4.9.3
python-dateutil~=2.8.1